util/amdfwtool: Use uint8_t for bitfields
Using a signed, non-fixed-width type for bitfields can cause problems.
So, use uint8_t since the affected bitfields occupy exactly one byte.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/91855
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ util/amdfwtool/amdfwtool.h
@@ -296,11 +296,11 @@ typedef struct _bios_directory_hdr {
 typedef struct _bios_directory_entry {
 	uint8_t type;
 	uint8_t region_type;
-	int reset:1;
-	int copy:1;
-	int ro:1;
-	int compressed:1;
-	int inst:4;
+	uint8_t reset:1;
+	uint8_t copy:1;
+	uint8_t ro:1;
+	uint8_t compressed:1;
+	uint8_t inst:4;
 	uint8_t subprog:3;
 	uint8_t romid:2;
 	uint8_t writable:1;
